Fix "should not sign in with wrong credentials"
The test "should not sign in with wrong credentials" in
test/controllers/sessions_controller_test.rb was looking for a redirect
to the sign in page when users who failed sign in were accessing the home
page. This assertion was failing because users are currently allowed to
access the home page (and most pages) even if they are not signed in.

To fix this, I checked if a user who failed sign in was being redirected
when trying to access the events edit page, which is currently only
accessible to signed in users.