Fix enhance_profile_job_test
The test "search Design a user with no name" was failing because it was
expecting to get "design" as the github account for the speaker "Design",
when in reality it would get "Design-and-Code". There are no users called
exactly "Design" on GitHub.

I tried deleting the cassette and re-generating it, and checked the
response. There are no users before Design-and-Code with a name containing
"design" as a separate word, and since their login is "Design-and-Code",
it seems the expected response should indeed be "Design-and-Code".