Attribute RailsConf 2013 "Crafting Gems" to correct Pat
This talk was given by Pat Allan, not Pat Allen. Almost certainly a typo on the
upload to youtube, and probably just luck that there is a patallen on github
that we could attribute it to.  We remove the "Pat Allen" speaker because
although they do exist their github profile doesn't suggest like they are a
rubyist - and with this data tweak, they have no talks anyway.