Merge Pan Thomakos and Panayiotis Thomakos
Both videos (Euruko 2018 & Goruco 2017) are by the same person who introduces
them selves as "Pan" not "Panayiotis" so let's choose the name they use.
Luckily the "Pan" version is also the one with extra detail in speakers.yml
anyway, so we set `canonical_slug` on the "Panayiotis" one to redirect to the
"Pan" one.