Tidy up Euruko 2016 Lightning talks
One of the speakers for day 1 lightning talks had an embedded special character
meaning the `# ...` talk title wasn't being treated as a comment, but a full
string. This meant we created two Melanie Keatleys - one for a talk at Euruko
2019 with her name and one for these lightning talks with the title of the talk
added to her name. We tidy up the list of speakers, and while we're at it turn
the comments in the speaker list into a better description. Gets rid of an
extra speaker _and_ provides some better descriptions. Win win!