Get rid of the "Yukihio 'Matz' Matsumoto" speaker
It's a typo and punctuation error for the "real" version: "Yukihiro "Matz"
Matsumoto". The Yukihio version isn't in speakers.yml, but there is both
"Yukihiro "Matz" Matsumoto" and "Yukihiro Matsumoto", but as they share a slug
it's unclear to me if that's an intentional redirect thing or not, so I've left
it.