Fix Andrew Shafer vs Andrew Schafer
Andrew gave two talks at MountainViewRuby - a lightning talk on puppet in 2008
and a full talk also on puppet in 2009. Although the videos are labelled with
Andrew Schafer and Andrew Shafer respectively, it's clear from the videos that
they're given by the same person, so lets add an entry in speakers.yml and
attribute both to one of them. Andrew Shafer appears to be the correct spelling
based on reference to github user in the 2009 talk.