Clarify password change side-effect at the controller level
Changing a password has ramifications for the system, I'd say that's part of the main story that the controller is telling and so should be included in the controller.

So we extract the logic into the domain model method `sign_out_siblings` on `Session`.

This also means we don't tie `Current` context to `User` in this case.